requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
idna==3.3
mysql-connector-python==8.0.28
oauthlib==3.2.0
phonenumbers==8.12.46
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python-slugify==6.1.1
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.9
